added pages in Glass
