+++ glass/p/opdrachten/handler.php
@@ -1,0 +1,8 @@
+<?php
+require_once "../../core/init.php";
+
+switch ($_GET['a']){
+    case 'check':
+        Database::update("assignments", array("done" => true), 'id', Database::escape($_GET["id"]));
+        break;
+}
